pyproject.toml
[tool.poetry]
name = "Your Porject Name"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
mlflow = "^1.2"
tensorflow = {version = "^1.14",optional = true}
tensorflow-gpu = {version = "^1.14",optional = true}
scikit-learn = "^0.21.3"
pandas = "^0.25.0"
pillow = "^6.1"
scipy = "^1.3"
numpy = "^1.17"
matplotlib = "^3.1"
seaborn = "^0.9.0"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
mypy = "^0.720.0"
pydocstyle = "^4.0"
flake8 = "^3.7"
autopep8 = "^1.4"
nose = "^1.3"

[tool.poetry.extras]
cpu = ["tensorflow"]
gpu = ["tensorflow-gpu"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
